refactor find invoiceable rewards
